refactor: split relatorios-notifications-service into smaller files
Split relatorios-notifications-service.ts (509 lines) into:
- relatorios-notifications-service.ts (218 lines) - Email notifications
- relatorios-finais-export-service.ts (288 lines) - Excel/ata generation

Both files now under 400-line max per CLAUDE.md guidelines.
